requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
PyMySQL==1.1.0
cryptography==41.0.3
python-dotenv==1.0.0